the ready commit for Hadoop usage
It requires a parameter of configuration: hipi.map.tasks which is
the number of map tasks you want to run since JobConf and related
setNumMapTasks has been deprecated. I attached a test example along
with the code along with build.xml. Compile with ant (assume /opt/hadoop
is your installation directory of hadoop). and with all these goodies
of Hadoop you are ready to go.

Before doing that, you can use the test image bundle, which usually
generated after you ran a JUnit test (/tmp/bundle.hib and
/tmp/bundle.hib.dat). Run the job with -D hipi.map.tasks=X argument.

The output is the camera model along with image width.

It requires more stress tests for current implementation.

git-svn-id: svn+ssh://galicia/usr/local/svn/HIPI@18 0c4b1d1f-a87e-43e2-b515-cbd7e906cf19